% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qc_visuals.R
\name{create_qc_visuals}
\alias{create_qc_visuals}
\title{Create tables and figures summarizing quality control data}
\usage{
create_qc_visuals(
  dataset,
  form = "coarse",
  summaryPeriod = "month",
  parameters_wo_blanks = c("DO", "pH", "Temperature", "Turbidity",
    "Specific conductivity", "Conductivity", "TDS"),
  field_only = TRUE,
  collection_rate_target = 10,
  replicate_variation_threshold = 20
)
}
\arguments{
\item{dataset}{list produced by summarizeQC}

\item{form}{type of table desired; 'coarse', 'blank', or 'rep'}

\item{summaryPeriod}{time period for summary plots; `all`, `year`, `month`, or `day`}

\item{parameters_wo_blanks}{parameters with no expectation for blanks or lab reps}

\item{field_only}{should blank table focus only on field blanks?}

\item{collection_rate_target}{target collection rate for quality control measures (blanks and replicates)}

\item{replicate_variation_threshold}{variation threshold used for evaluating replicate performance: what is the target level of precision?}
}
\value{
a table for inclusion in markdown documents
}
\description{
Create tables and figures summarizing quality control data
}
