% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Generate a Water Quality Report for an organization}
\usage{
create_report(
  org = "TURTLEMT",
  startDate = "01-01-2015",
  endDate = "12-31-2022",
  parameters = 1:11,
  extFile = "script_generateReport.qmd",
  prompt_user = TRUE,
  output = "html",
  output_directory = file.path(getwd(), "R8WD_output"),
  draft_report = TRUE
)
}
\arguments{
\item{org}{organization (short name in WQP) whose data should be used to build the report. Multiple organization names can be input as a character vector; names will be homogenized and data will be pooled for analysis purposes.}

\item{startDate}{start date for data used in the report, in format `\%m-\%d-\%Y`}

\item{endDate}{final date for data used in the report, in format `\%m-\%d-\%Y`}

\item{parameters}{parameters to use in report. Must be entered as an index of acceptable parameters listed in params$params (e.g., `parameters = c(1:4,7)`)}

\item{extFile}{name of report-generating script, located in the inst/extdata folder of the R8WD R package}

\item{prompt_user}{if TRUE, user is prompted to use one of the organization codes in the `tribes` object provided with `R8WD`.}

\item{output}{either 'docx' or 'html'}

\item{output_directory}{directory where output will be located. If the directory doesn't exist it will be created.}

\item{draft_report}{TRUE or FALSE}
}
\value{
Quarto markdown document
}
\description{
Generate a Water Quality Report for an organization
}
\examples{
create_report(org = 'Tribe 1')
create_report(org = 'Tribe 2')

}
