% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_tables.R
\docType{data}
\name{criteria_tables}
\alias{criteria_tables}
\title{Data: Aquatic and Human Health Criteria Table}
\format{
An rda file with a list of two data frames of various water quality data
\itemize{
  \item aq_life_criteria a data frame of aquatic life criteria
  \item e_coli_criteria a data frame for e coli criteria
}
}
\usage{
criteria_tables
}
\description{
Data: Aquatic and Human Health Criteria Table
}
\keyword{criteria}
