% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate Coefficient of Variation (as percent not decimal fraction, e.g., values fall on a scale of 0-100)}
\usage{
cv(x)
}
\arguments{
\item{x}{a vector of two or more values}
}
\value{
a single numeric value
}
\description{
Calculate Coefficient of Variation (as percent not decimal fraction, e.g., values fall on a scale of 0-100)
}
\examples{
cv(c(10, 20))
cv(c(10:30))
}
