% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pa.R
\name{get_pa}
\alias{get_pa}
\title{Get Purple Air data from Purple Air API (based on https://api.purpleair.com/)}
\usage{
get_pa(
  api_key,
  start_date = as.POSIXct("2023-05-01", format = "\%Y-\%m-\%d"),
  end_date = as.POSIXct(Sys.time()),
  sensor = 2478,
  average_interval = 1440,
  parameters = c("humidity", "temperature", "pressure", "pm2.5_atm_a", "pm2.5_atm_b")
)
}
\arguments{
\item{api_key}{an API key obtained from your personal Purple Air account. Your first 1m tokens (i.e., data points downloaded) are free.}

\item{start_date}{start date for period of record}

\item{end_date}{end date for period of record}

\item{sensor}{sensor number. This is assigned by purple air and can be identified using their sensor map}

\item{average_interval}{averaging interval for returned data, expressed in minutes. The acceptable inputs are 0 (real-time), 10 (default if not specified), 30, 60, 360 (6 hour), 1440 (1 day)}

\item{parameters}{vector of parameter names to download. For a complete list of fields, see https://api.purpleair.com/}
}
\value{
dataframe
}
\description{
Get Purple Air data from Purple Air API (based on https://api.purpleair.com/)
}
