% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinWQPProfiles.R
\name{joinWQPProfiles}
\alias{joinWQPProfiles}
\title{Join WQP Measurement and Site Data}
\usage{
joinWQPProfiles(full = "null", sites = "null")
}
\arguments{
\item{dat.full}{output of query to WQP portal using readWQPdata function in
the dataRetrieval package}

\item{dat.sites}{output of query to WQP portal using whatWQPdat.sites function in the
dataRetrieval package}
}
\value{
a combined data frame of WQP measurements with corresponding sampling
        site data
}
\description{
This function takes data from the same WQP query and combines
             site data with measurements to map observations with physical
             sampling locations
}
