% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_raw_data.R
\name{merge_raw_data}
\alias{merge_raw_data}
\title{Merge raw Purple Air .csv files downloaded from SD card storage}
\usage{
merge_raw_data(directory, output_file = NULL, replace_file = TRUE)
}
\arguments{
\item{directory}{Directory with csv files containing daily data. Follow R syntax - use '/' or '\\' to separate directories (e.g., `directory = "CCST/CAA105/data"`).}

\item{output_file}{Name of output data file, including extension (e.g., `final_data.csv`). Optional; if left as NULL the code will generate a date-stamped .csv file in `directory`. If this argument is provided, the file must be a csv.}

\item{replace_file}{If an output file is provided, this argument indicates whether to overwrite existing data, or append new data to the file.}
}
\value{
a dataframe is returned to the R console, and a merged csv file is written to `directory`
}
\description{
Merge raw Purple Air .csv files downloaded from SD card storage
}
