% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_location_types.R
\docType{data}
\name{monitoring_location_types}
\alias{monitoring_location_types}
\title{Data: Monitoring Location Types with Associated Map Marker Label}
\format{
A dataframe with Water Quality Portal monitoring location types, to be used in reassigning simplified values for custom data markers in the create_map function
\itemize{
  \item Name WQP monitoring location types
  \item Label simplified name for map markers
}
}
\usage{
monitoring_location_types
}
\description{
Data: Monitoring Location Types with Associated Map Marker Label
}
\keyword{wqp}
