% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{params}
\alias{params}
\title{Data: Core parameters}
\format{
A dataframe with Water Quality Portal characteristic names for core water quality parameters, allowing re-naming of duplicated parameters.
\itemize{
  \item params WQP characteristic names
  \item new_param pooled parameter categories
}
}
\usage{
params
}
\description{
Data: Core parameters
}
\examples{
params
# tribeData2 <- preProcessResults(tribeData1, multiplier = 0.5)
# tribeData2$CharacteristicName <-
#  params$new_param[match(tribeData2$CharacteristicName, params$params)]

}
\keyword{wqp}
