% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessResults.R
\name{preProcessResults}
\alias{preProcessResults}
\title{Process raw WQP data prior to analysis}
\usage{
preProcessResults(
  data,
  multiplier = 0.5,
  value_column = "ResultMeasureValue",
  convert_ug_to_mg = TRUE,
  DL_column = "DetectionQuantitationLimitMeasure.MeasureValue",
  unit_column = "ResultMeasure.MeasureUnitCode"
)
}
\arguments{
\item{data}{data downloaded from WQP}

\item{multiplier}{multiplier applied to the detection limit when replacing values below detection limits. Default is 0.5}

\item{value_column}{name of the column in `data` containing values/concentrations}

\item{convert_ug_to_mg}{logical; if TRUE, values with units of ug/L are converted to mg/L}

\item{DL_column}{name of the column containing detection limit information}

\item{unit_column}{name of column identifying units}
}
\value{
dataframe
}
\description{
Process raw WQP data prior to analysis
}
