% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculate root mean squared error}
\usage{
rmse(observed, predicted)
}
\arguments{
\item{observed}{vector of observed values}

\item{predicted}{vector of predicted values}
}
\value{
a vector of numeric values
}
\description{
Calculate root mean squared error
}
\examples{
rmse(1:10, 11:20)
}
