% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpd.R
\name{rpd}
\alias{rpd}
\title{Calculate Relative Percent Difference}
\usage{
rpd(x)
}
\arguments{
\item{x}{a vector of two values}
}
\value{
a single numeric value
}
\description{
Calculate Relative Percent Difference
}
\examples{
rpd(c(10, 20))
}
