% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentencify.R
\name{sentencify}
\alias{sentencify}
\title{Convert a vector into a grammatically correct character element}
\usage{
sentencify(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a character element
}
\description{
Convert a vector into a grammatically correct character element
}
\examples{
sentencify(1); sentencify(1:2); sentencify(1:5)
}
