% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Spider Monkeys}
\usage{
test_function(x)
}
\arguments{
\item{x}{}
}
\value{
single chr element
}
\description{
Spider Monkeys
}
\examples{
test_function("chimpanzee")
}
