% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribes.R
\docType{data}
\name{tribes}
\alias{tribes}
\title{Data: Tribal organization names}
\format{
A vector of Water Quality Portal organization names for Region 8 Tribes.
\itemize{
  \item tribes a vector of Tribal organization names
}
}
\usage{
tribes
}
\description{
Data: Tribal organization names
}
\examples{
tribes

}
\keyword{wqp}
